// ai-products-selection-frontend/src/app/core/api/model/weightVersionDetail.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SceneWeightSet } from './sceneWeightSet';


export interface WeightVersionDetail { 
    id?: number;
    versionNo?: string;
    name?: string;
    status?: WeightVersionDetail.StatusEnum;
    effectiveFrom?: string;
    current?: boolean;
    changeNote?: string;
    scenes?: Array<SceneWeightSet>;
}
export namespace WeightVersionDetail {
    export type StatusEnum = 'DRAFT' | 'APPROVED' | 'RETIRED';
    export const StatusEnum = {
        Draft: 'DRAFT' as StatusEnum,
        Approved: 'APPROVED' as StatusEnum,
        Retired: 'RETIRED' as StatusEnum
    };
}