// ai-products-selection-frontend/src/app/core/api/model/sceneWeightSet.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SceneWeightSet { 
    sceneType: SceneWeightSet.SceneTypeEnum;
    weights: { [key: string]: number; };
    gradeAMin: number;
    gradeBMin: number;
}
export namespace SceneWeightSet {
    export type SceneTypeEnum = 'VIRAL' | 'FESTIVAL' | 'REPLENISHMENT' | 'SEASONAL';
    export const SceneTypeEnum = {
        Viral: 'VIRAL' as SceneTypeEnum,
        Festival: 'FESTIVAL' as SceneTypeEnum,
        Replenishment: 'REPLENISHMENT' as SceneTypeEnum,
        Seasonal: 'SEASONAL' as SceneTypeEnum
    };
}